--- src/main/java/com/profect/tickle/global/config/CorsConfig.java
@@ -1,23 +1,0 @@
-package com.profect.tickle.global.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class CorsConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-                .allowedOrigins(
-                        "http://localhost:5173",  // Vite React 개발 서버
-                        "http://localhost:3000",  // Create React App (필요시)
-                        "http://localhost:8081"   // 기타 필요한 포트
-                )
-                .allowedMethods("GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS")
-                .allowedHeaders("*")
-                .allowCredentials(true)
-                .maxAge(3600); // preflight 캐시 시간 (1시간)
-    }
-}
