+++ src/main/java/com/profect/tickle/global/config/CorsConfigLocal.java
@@ -1,0 +1,24 @@
+package com.profect.tickle.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@Profile("local")
+public class CorsConfigLocal implements WebMvcConfigurer {
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/api/**")
+                .allowedOrigins(
+                        "http://localhost:5173",   // Vite dev
+                        "http://localhost:3000",    // CRA dev (필요시)
+                        "http://localhost:8081"
+                )
+                .allowedMethods("GET","POST","PUT","PATCH","DELETE","OPTIONS")
+                .allowedHeaders("*")
+                .allowCredentials(true)
+                .maxAge(3600);
+    }
+}
