+++ src/main/java/com/profect/tickle/global/config/CorsConfigProd.java
@@ -1,0 +1,19 @@
+package com.profect.tickle.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@Profile("prod")
+public class CorsConfigProd implements WebMvcConfigurer {
+    @Override public void addCorsMappings(CorsRegistry r) {
+        r.addMapping("/api/**")
+                .allowedOrigins("https://tickle.kr","https://www.tickle.kr")
+                .allowedMethods("GET","POST","PUT","PATCH","DELETE","OPTIONS")
+                .allowedHeaders("*")
+                .allowCredentials(true) // 쿠키 인증이면 true
+                .maxAge(3600);
+    }
+}
